Split tests into separate files and targets
