+++ test/CMakeLists.txt
@@ -5,25 +5,38 @@ cmake_minimum_required(VERSION 3.8.2)
 
 find_package(GTest MODULE REQUIRED)
 
-# tests
-add_executable(tests
-  tests.cpp)
-target_link_libraries(tests PRIVATE
+add_executable(today_tests TodayTests.cpp)
+target_link_libraries(today_tests PRIVATE
   todaygraphql
   graphqljson
   GTest::GTest
   GTest::Main)
+gtest_add_tests(TARGET today_tests)
 
-gtest_add_tests(TARGET tests)
-#add_test(NAME TodayServiceCase
-#  COMMAND tests --gtest_filter=TodayServiceCase.*
-#  WORKING_DIRECTORY $<TARGET_FILE_DIR:tests>)
-#add_test(NAME ArgumentsCase
-#  COMMAND tests --gtest_filter=ArgumentsCase.*
-#  WORKING_DIRECTORY $<TARGET_FILE_DIR:tests>)
-#add_test(NAME PegtlCase
-#  COMMAND tests --gtest_filter=PegtlCase.*
-#  WORKING_DIRECTORY $<TARGET_FILE_DIR:tests>)
-#add_test(NAME ResponseCase
-#  COMMAND tests --gtest_filter=ResponseCase.*
-#  WORKING_DIRECTORY $<TARGET_FILE_DIR:tests>)
+add_executable(argument_tests ArgumentTests.cpp)
+target_link_libraries(argument_tests PRIVATE
+  todaygraphql
+  graphqljson
+  GTest::GTest
+  GTest::Main)
+gtest_add_tests(TARGET argument_tests)
+
+add_executable(pegtl_tests PegtlTests.cpp)
+target_link_libraries(pegtl_tests PRIVATE
+  GTest::GTest
+  GTest::Main)
+target_include_directories(pegtl_tests PUBLIC
+  ${CMAKE_BINARY_DIR}/include
+  ${CMAKE_SOURCE_DIR}/include
+  ${CMAKE_SOURCE_DIR}/PEGTL/include)
+gtest_add_tests(TARGET pegtl_tests)
+
+add_executable(response_tests ResponseTests.cpp)
+target_link_libraries(response_tests PRIVATE
+  graphqlservice
+  GTest::GTest
+  GTest::Main)
+target_include_directories(response_tests PUBLIC
+  ${CMAKE_BINARY_DIR}/include
+  ${CMAKE_SOURCE_DIR}/include)
+gtest_add_tests(TARGET response_tests)
