+++ test/TodayTests.cpp
@@ -4,18 +4,12 @@
 #include <gtest/gtest.h>
 
 #include "Today.h"
-#include "GraphQLGrammar.h"
 
 #include <graphqlservice/JSONResponse.h>
-
-#include <tao/pegtl/analyze.hpp>
 
 #include <chrono>
 
 using namespace facebook::graphql;
-using namespace facebook::graphql::peg;
-
-using namespace tao::graphqlpeg;
 
 using namespace std::literals;
 
@@ -1220,598 +1214,4 @@ TEST_F(TodayServiceCase, NonBlockingQueryAppointmentsById)
 	{
 		FAIL() << response::toJSON(response::Value(ex.getErrors()));
 	}
-}
-
-TEST(ArgumentsCase, ListArgumentStrings)
-{
-	auto parsed = response::parseJSON(R"js({"value":[
-		"string1",
-		"string2",
-		"string3"
-	]})js");
-	std::vector<std::string> actual;
-
-	try
-	{
-		actual = service::StringArgument::require<service::TypeModifier::List>("value", parsed);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	ASSERT_EQ(3, actual.size()) << "should get 3 entries";
-	EXPECT_EQ("string1", actual[0]) << "entry should match";
-	EXPECT_EQ("string2", actual[1]) << "entry should match";
-	EXPECT_EQ("string3", actual[2]) << "entry should match";
-}
-
-TEST(ArgumentsCase, ListArgumentStringsNonNullable)
-{
-	auto parsed = response::parseJSON(R"js({"value":[
-		"string1",
-		null,
-		"string2",
-		"string3"
-	]})js");
-	bool caughtException = false;
-	std::string exceptionWhat;
-
-	try
-	{
-		auto actual = service::StringArgument::require<service::TypeModifier::List>("value", parsed);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		exceptionWhat = response::toJSON(response::Value(ex.getErrors()));
-		caughtException = true;
-	}
-
-	ASSERT_TRUE(caughtException);
-	EXPECT_EQ(R"js([{"message":"Invalid argument: value error: not a string"}])js", exceptionWhat) << "exception should match";
-}
-
-TEST(ArgumentsCase, ListArgumentStringsNullable)
-{
-	auto parsed = response::parseJSON(R"js({"value":[
-		"string1",
-		"string2",
-		null,
-		"string3"
-	]})js");
-	std::vector<std::optional<std::string>> actual;
-
-	try
-	{
-		actual = service::StringArgument::require<
-			service::TypeModifier::List,
-			service::TypeModifier::Nullable
-		>("value", parsed);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	ASSERT_EQ(4, actual.size()) << "should get 4 entries";
-	ASSERT_TRUE(actual[0].has_value()) << "should not be null";
-	EXPECT_EQ("string1", *actual[0]) << "entry should match";
-	ASSERT_TRUE(actual[1].has_value()) << "should not be null";
-	EXPECT_EQ("string2", *actual[1]) << "entry should match";
-	EXPECT_FALSE(actual[2].has_value()) << "should be null";
-	ASSERT_TRUE(actual[3].has_value()) << "should not be null";
-	EXPECT_EQ("string3", *actual[3]) << "entry should match";
-}
-
-TEST(ArgumentsCase, ListArgumentListArgumentStrings)
-{
-	auto parsed = response::parseJSON(R"js({"value":[
-		["list1string1", "list1string2"],
-		["list2string1", "list2string2"]
-	]})js");
-	std::vector<std::vector<std::string>> actual;
-
-	try
-	{
-		actual = service::StringArgument::require<
-			service::TypeModifier::List,
-			service::TypeModifier::List
-		>("value", parsed);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	ASSERT_EQ(2, actual.size()) << "should get 2 entries";
-	ASSERT_EQ(2, actual[0].size()) << "should get 2 entries";
-	EXPECT_EQ("list1string1", actual[0][0]) << "entry should match";
-	EXPECT_EQ("list1string2", actual[0][1]) << "entry should match";
-	ASSERT_EQ(2, actual[1].size()) << "should get 2 entries";
-	EXPECT_EQ("list2string1", actual[1][0]) << "entry should match";
-	EXPECT_EQ("list2string2", actual[1][1]) << "entry should match";
-}
-
-TEST(ArgumentsCase, ListArgumentNullableListArgumentStrings)
-{
-	auto parsed = response::parseJSON(R"js({"value":[
-		null,
-		["list2string1", "list2string2"]
-	]})js");
-	std::vector<std::optional<std::vector<std::string>>> actual;
-
-	try
-	{
-		actual = service::StringArgument::require<
-			service::TypeModifier::List,
-			service::TypeModifier::Nullable,
-			service::TypeModifier::List
-		>("value", parsed);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	ASSERT_EQ(2, actual.size()) << "should get 2 entries";
-	EXPECT_FALSE(actual[0].has_value()) << "should be null";
-	ASSERT_EQ(2, actual[1]->size()) << "should get 2 entries";
-	EXPECT_EQ("list2string1", (*actual[1])[0]) << "entry should match";
-	EXPECT_EQ("list2string2", (*actual[1])[1]) << "entry should match";
-}
-
-TEST(ArgumentsCase, TaskStateEnum)
-{
-	response::Value response(response::Type::Map);
-	response::Value status(response::Type::EnumValue);
-	status.set<response::StringType>("Started");
-	response.emplace_back("status", std::move(status));
-	today::TaskState actual = static_cast<today::TaskState>(-1);
-
-	try
-	{
-		actual = service::ModifiedArgument<today::TaskState>::require("status", response);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	EXPECT_EQ(today::TaskState::Started, actual) << "should parse the enum";
-}
-
-TEST(ArgumentsCase, TaskStateEnumFromString)
-{
-	response::Value response(response::Type::Map);
-	response::Value status("Started");
-	response.emplace_back("status", std::move(status));
-	today::TaskState actual = static_cast<today::TaskState>(-1);
-	bool caughtException = false;
-	std::string exceptionWhat;
-
-	try
-	{
-		actual = service::ModifiedArgument<today::TaskState>::require("status", response);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		caughtException = true;
-		exceptionWhat = response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	EXPECT_NE(today::TaskState::Started, actual) << "should not parse the enum from a known string value";
-	ASSERT_TRUE(caughtException);
-	EXPECT_EQ(R"js([{"message":"Invalid argument: status error: not a valid TaskState value"}])js", exceptionWhat) << "exception should match";
-}
-
-TEST(ArgumentsCase, TaskStateEnumFromJSONString)
-{
-	response::Value response(response::Type::Map);
-	response::Value status("Started");
-	response.emplace_back("status", status.from_json());
-	today::TaskState actual = static_cast<today::TaskState>(-1);
-
-
-	try
-	{
-		actual = service::ModifiedArgument<today::TaskState>::require("status", response);
-	}
-	catch (const service::schema_exception& ex)
-	{
-		FAIL() << response::toJSON(response::Value(ex.getErrors()));
-	}
-
-	EXPECT_EQ(today::TaskState::Started, actual) << "should parse the enum";
-}
-
-TEST(PegtlCase, ParseKitchenSinkQuery)
-{
-	memory_input<> input(R"gql(
-		# Copyright (c) 2015-present, Facebook, Inc.
-		#
-		# This source code is licensed under the MIT license found in the
-		# LICENSE file in the root directory of this source tree.
-
-		query queryName($foo: ComplexType, $site: Site = MOBILE) {
-		  whoever123is: node(id: [123, 456]) {
-			id ,
-			... on User @defer {
-			  field2 {
-				id ,
-				alias: field1(first:10, after:$foo,) @include(if: $foo) {
-				  id,
-				  ...frag
-				}
-			  }
-			}
-			... @skip(unless: $foo) {
-			  id
-			}
-			... {
-			  id
-			}
-		  }
-		}
-
-		mutation likeStory {
-		  like(story: 123) @defer {
-			story {
-			  id
-			}
-		  }
-		}
-
-		subscription StoryLikeSubscription($input: StoryLikeSubscribeInput) {
-		  storyLikeSubscribe(input: $input) {
-			story {
-			  likers {
-				count
-			  }
-			  likeSentence {
-				text
-			  }
-			}
-		  }
-		}
-
-		fragment frag on Friend {
-		  foo(size: $size, bar: $b, obj: {key: "value", block: """
-
-			  block string uses \"""
-
-		  """})
-		}
-
-		{
-		  unnamed(truthy: true, falsey: false, nullish: null),
-		  query
-		})gql", "ParseKitchenSinkQuery");
-
-	const bool result = parse<document>(input);
-
-	ASSERT_TRUE(result) << "we should be able to parse the doc";
-}
-
-TEST(PegtlCase, ParseKitchenSinkSchema)
-{
-	memory_input<> input(R"gql(
-		# Copyright (c) 2015-present, Facebook, Inc.
-		#
-		# This source code is licensed under the MIT license found in the
-		# LICENSE file in the root directory of this source tree.
-
-		# (this line is padding to maintain test line numbers)
-
-		schema {
-		  query: QueryType
-		  mutation: MutationType
-		}
-
-		type Foo implements Bar {
-		  one: Type
-		  two(argument: InputType!): Type
-		  three(argument: InputType, other: String): Int
-		  four(argument: String = "string"): String
-		  five(argument: [String] = ["string", "string"]): String
-		  six(argument: InputType = {key: "value"}): Type
-		  seven(argument: Int = null): Type
-		}
-
-		type AnnotatedObject @onObject(arg: "value") {
-		  annotatedField(arg: Type = "default" @onArg): Type @onField
-		}
-
-		interface Bar {
-		  one: Type
-		  four(argument: String = "string"): String
-		}
-
-		interface AnnotatedInterface @onInterface {
-		  annotatedField(arg: Type @onArg): Type @onField
-		}
-
-		union Feed = Story | Article | Advert
-
-		union AnnotatedUnion @onUnion = A | B
-
-		scalar CustomScalar
-
-		scalar AnnotatedScalar @onScalar
-
-		enum Site {
-		  DESKTOP
-		  MOBILE
-		}
-
-		enum AnnotatedEnum @onEnum {
-		  ANNOTATED_VALUE @onEnumValue
-		  OTHER_VALUE
-		}
-
-		input InputType {
-		  key: String!
-		  answer: Int = 42
-		}
-
-		input AnnotatedInput @onInputObjectType {
-		  annotatedField: Type @onField
-		}
-
-		extend type Foo {
-		  seven(argument: [String]): Type
-		}
-
-		# NOTE: out-of-spec test cases commented out until the spec is clarified; see
-		# https://github.com/graphql/graphql-js/issues/650 .
-		# extend type Foo @onType {}
-
-		#type NoFields {}
-
-		directive @skip(if: Boolean!) on FIELD | FRAGMENT_SPREAD | INLINE_FRAGMENT
-
-		directive @include(if: Boolean!)
-		  on FIELD
-		   | FRAGMENT_SPREAD
-		   | INLINE_FRAGMENT)gql", "ParseKitchenSinkSchema");
-
-	const bool result = parse<document>(input);
-
-	ASSERT_TRUE(result) << "we should be able to parse the doc";
-}
-
-TEST(PegtlCase, ParseKitchenSink)
-{
-	memory_input<> input(R"gql(
-		# Copyright (c) 2015-present, Facebook, Inc.
-		#
-		# This source code is licensed under the MIT license found in the
-		# LICENSE file in the root directory of this source tree.
-
-		query queryName($foo: ComplexType, $site: Site = MOBILE) {
-		  whoever123is: node(id: [123, 456]) {
-			id ,
-			... on User @defer {
-			  field2 {
-				id ,
-				alias: field1(first:10, after:$foo,) @include(if: $foo) {
-				  id,
-				  ...frag
-				}
-			  }
-			}
-			... @skip(unless: $foo) {
-			  id
-			}
-			... {
-			  id
-			}
-		  }
-		}
-
-		mutation likeStory {
-		  like(story: 123) @defer {
-			story {
-			  id
-			}
-		  }
-		}
-
-		subscription StoryLikeSubscription($input: StoryLikeSubscribeInput) {
-		  storyLikeSubscribe(input: $input) {
-			story {
-			  likers {
-				count
-			  }
-			  likeSentence {
-				text
-			  }
-			}
-		  }
-		}
-
-		fragment frag on Friend {
-		  foo(size: $size, bar: $b, obj: {key: "value", block: """
-
-			  block string uses \"""
-
-		  """})
-		}
-
-		{
-		  unnamed(truthy: true, falsey: false, nullish: null),
-		  query
-		})gql", "ParseKitchenSinkSchema");
-
-	const bool result = parse<document>(input);
-
-	ASSERT_TRUE(result) << "we should be able to parse the doc";
-}
-
-TEST(PegtlCase, ParseTodayQuery)
-{
-	memory_input<> input(R"gql(
-		query Everything {
-			appointments {
-				edges {
-					node {
-						id
-						subject
-						when
-						isNow
-					}
-				}
-			}
-			tasks {
-				edges {
-					node {
-						id
-						title
-						isComplete
-					}
-				}
-			}
-			unreadCounts {
-				edges {
-					node {
-						id
-						name
-						unreadCount
-					}
-				}
-			}
-		})gql", "ParseTodayQuery");
-
-	const bool result = parse<document>(input);
-
-	ASSERT_TRUE(result) << "we should be able to parse the doc";
-}
-
-TEST(PegtlCase, ParseTodaySchema)
-{
-	memory_input<> input(R"gql(
-		# Copyright (c) Microsoft Corporation. All rights reserved.
-		# Licensed under the MIT License.
-
-		schema {
-			query: Query
-			mutation: Mutation
-			subscription: Subscription
-		}
-
-		scalar ItemCursor
-
-		type Query {
-			node(id: ID!) : Node
-
-			appointments(first: Int, after: ItemCursor, last: Int, before: ItemCursor): AppointmentConnection!
-			tasks(first: Int, after: ItemCursor, last: Int, before: ItemCursor): TaskConnection!
-			unreadCounts(first: Int, after: ItemCursor, last: Int, before: ItemCursor): FolderConnection!
-
-			appointmentsById(ids: [ID!]!) : [Appointment]!
-			tasksById(ids: [ID!]!): [Task]!
-			unreadCountsById(ids: [ID!]!): [Folder]!
-		}
-
-		interface Node {
-			id: ID!
-		}
-
-		type PageInfo {
-			hasNextPage: Boolean!
-			hasPreviousPage: Boolean!
-		}
-
-		type AppointmentEdge {
-			node: Appointment
-			cursor: ItemCursor!
-		}
-
-		type AppointmentConnection {
-			pageInfo: PageInfo!
-			edges: [AppointmentEdge]
-		}
-
-		type TaskEdge {
-			node: Task
-			cursor: ItemCursor!
-		}
-
-		type TaskConnection {
-			pageInfo: PageInfo!
-			edges: [TaskEdge]
-		}
-
-		type FolderEdge {
-			node: Folder
-			cursor: ItemCursor!
-		}
-
-		type FolderConnection {
-			pageInfo: PageInfo!
-			edges: [FolderEdge]
-		}
-
-		input CompleteTaskInput {
-			id: ID!
-			isComplete: Boolean = true
-			clientMutationId: String
-		}
-
-		type CompleteTaskPayload {
-			task: Task
-			clientMutationId: String
-		}
-
-		type Mutation {
-			completeTask(input: CompleteTaskInput!) : CompleteTaskPayload!
-		}
-
-		type Subscription {
-			nextAppointmentChange : Appointment
-		}
-
-		scalar DateTime
-
-		enum TaskState {
-			New
-			Started
-			Complete
-		}
-
-		type Appointment implements Node {
-			id: ID!
-			when: DateTime
-			subject: String
-			isNow: Boolean!
-		}
-
-		type Task implements Node {
-			id: ID!
-			title: String
-			isComplete: Boolean!
-		}
-
-		type Folder implements Node {
-			id: ID!
-			name: String
-			unreadCount: Int!
-		})gql", "ParseTodaySchema");
-
-	const bool result = parse<document>(input);
-
-	ASSERT_TRUE(result) << "we should be able to parse the doc";
-}
-
-TEST(PegtlCase, AnalyzeGrammar)
-{
-	ASSERT_EQ(0, analyze<document>(true)) << "there shuldn't be any infinite loops in the PEG version of the grammar";
-}
-
-TEST(ResponseCase, ValueConstructorFromStringLiteral)
-{
-	auto expected = "Test String";
-	auto actual = response::Value(expected);
-
-	ASSERT_TRUE(response::Type::String == actual.type());
-	ASSERT_EQ(expected, actual.release<response::StringType>());
 }
